requirements.txt
streamlit==1.34.0
torch==2.2.0
torchvision==0.18.0
pillow==10.0.0
matplotlib==3.8.0
scikit-learn==1.4.0
tqdm==4.66.1
albumentations==1.3.0
pandas==2.2.0
